requirements.txt
flask==2.0.1
flask-cors==3.0.10
psycopg2-binary==2.9.1
line-bot-sdk==2.4.2
python-dotenv==0.19.0
werkzeug==2.0.1 